Protocol fee sweeper V2 deployment preparation.
