+++ v3/deprecated/20250228-v3-protocol-fee-sweeper/output/arbitrum.json
@@ -1,0 +1,3 @@
+{
+  "ProtocolFeeSweeper": "0x3BEb058DE1A25dd24223fd9e1796df8589429AcE"
+}
