+++ v3/deprecated/20250228-v3-protocol-fee-sweeper/output/avalanche.json
@@ -1,0 +1,3 @@
+{
+  "ProtocolFeeSweeper": "0x5B42eC6D40f7B7965BE5308c70e2603c0281C1E9"
+}
