+++ v3/deprecated/20250228-v3-protocol-fee-sweeper/output/base.json
@@ -1,0 +1,3 @@
+{
+  "ProtocolFeeSweeper": "0x83E443EF4f9963C77bd860f94500075556668cb8"
+}
