+++ v3/deprecated/20250228-v3-protocol-fee-sweeper/output/gnosis.json
@@ -1,0 +1,3 @@
+{
+  "ProtocolFeeSweeper": "0x1bcA39b01F451b0a05D7030e6e6981a73B716b1C"
+}
