+++ v3/deprecated/20250228-v3-protocol-fee-sweeper/output/mainnet.json
@@ -1,0 +1,3 @@
+{
+  "ProtocolFeeSweeper": "0xd22eecBB495380Ef52b1CCeF1cA594979885D484"
+}
