+++ v3/deprecated/20250228-v3-protocol-fee-sweeper/output/sepolia.json
@@ -1,0 +1,3 @@
+{
+  "ProtocolFeeSweeper": "0x5E69888b605a181A5615aC1b3C8937aFA2913d15"
+}
