--- v3/tasks/20250228-v3-protocol-fee-sweeper/index.ts
@@ -1,9 +1,0 @@
-import { ProtocolFeeSweeperDeployment } from './input';
-import { Task, TaskRunOptions } from '@src';
-
-/* eslint-disable @typescript-eslint/no-non-null-assertion */
-export default async (task: Task, { force, from }: TaskRunOptions = {}): Promise<void> => {
-  const input = task.input() as ProtocolFeeSweeperDeployment;
-
-  await task.deployAndVerify('ProtocolFeeSweeper', [input.Vault, input.FeeRecipient], from, force);
-};
