--- v3/tasks/20250228-v3-protocol-fee-sweeper/output/arbitrum.json
@@ -1,3 +1,0 @@
-{
-  "ProtocolFeeSweeper": "0x3BEb058DE1A25dd24223fd9e1796df8589429AcE"
-}
