--- v3/tasks/20250228-v3-protocol-fee-sweeper/output/base.json
@@ -1,3 +1,0 @@
-{
-  "ProtocolFeeSweeper": "0x83E443EF4f9963C77bd860f94500075556668cb8"
-}
