--- v3/tasks/20250228-v3-protocol-fee-sweeper/output/gnosis.json
@@ -1,3 +1,0 @@
-{
-  "ProtocolFeeSweeper": "0x1bcA39b01F451b0a05D7030e6e6981a73B716b1C"
-}
