--- v3/tasks/20250228-v3-protocol-fee-sweeper/output/mainnet.json
@@ -1,3 +1,0 @@
-{
-  "ProtocolFeeSweeper": "0xd22eecBB495380Ef52b1CCeF1cA594979885D484"
-}
