--- v3/tasks/20250228-v3-protocol-fee-sweeper/output/optimism.json
@@ -1,3 +1,0 @@
-{
-  "ProtocolFeeSweeper": "0x86e67E115f96DF37239E0479441303De0de7bc2b"
-}
