--- v3/tasks/20250228-v3-protocol-fee-sweeper/output/sepolia.json
@@ -1,3 +1,0 @@
-{
-  "ProtocolFeeSweeper": "0x5E69888b605a181A5615aC1b3C8937aFA2913d15"
-}
